Project started

+++ Project/components/redis.sh
@@ -4,7 +4,9 @@ component=mongodb
 os_prereq
 
 print "Set up redis repo"
+if [ ! -f /etc/yum.repos.d/remi.repo ]; then
 yum install http://rpms.remirepo.net/enterprise/remi-release-7.rpm -y
+fi
 STAT $? "Setting up redis repo"
 
 print "Install redis"
